Move _is_geographic_crs to module where it's used.
