+++ harmony_regridding_service/message_utilities.py
@@ -2,8 +2,12 @@
 
 from harmony_service_lib.message import Message
 from harmony_service_lib.message_utility import rgetattr
+from pyproj import CRS
+from pyproj.exceptions import CRSError
 
-from harmony_regridding_service.crs import is_geographic_crs
+from harmony_regridding_service.exceptions import (
+    InvalidTargetCRS,
+)
 
 VALID_INTERPOLATION_METHODS = ('Elliptical Weighted Averaging',)
 
@@ -69,3 +73,18 @@ def has_valid_crs(message: Message) -> bool:
     """
     target_crs = rgetattr(message, 'format.crs')
     return target_crs is None or is_geographic_crs(target_crs)
+
+
+def is_geographic_crs(crs_string: str) -> bool:
+    """Infer if CRS is geographic.
+
+    Use pyproj to ascertain if the supplied Coordinate Reference System
+    (CRS) is geographic.
+    """
+    try:
+        crs = CRS(crs_string)
+        is_geographic = crs.is_geographic
+    except CRSError as exception:
+        raise InvalidTargetCRS(crs_string) from exception
+
+    return is_geographic
